pyproject.toml
[project]
name = "rag-book-kb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "beautifulsoup4==4.12.3",
    "chromadb==0.5.3",
    "ebooklib==0.18",
    "fastapi==0.111.0",
    "google-genai>=1.2.0",
    "google-generativeai==0.7.2",
    "groq==0.9.0",
    "httpx==0.27.0",
    "pydantic==2.7.4",
    "pymupdf==1.24.5",
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "uvicorn[standard]==0.30.1",
    "watchdog==4.0.1",
]
